012525 morning

+++ WebApp/lib/Analysis.jl
@@ -291,11 +291,11 @@ function timelapse_processing(images, blockDiameter, ntimepoints, shift_thresh, 
             else
                 OD_images[:,:,t] = @views (-1 .* log10.((images[:,:,t] .- Imin) ./ (images[:,:,1] .- Imin)))
             end
-            @inbounds biomasses[t] = @views mean(OD_images[:,:,t] .* masks[:,:,t])
+            @inbounds biomasses[t] = @views Float64(mean(OD_images[:,:,t] .* masks[:,:,t]))
         end
     else
         for t in 1:ntimepoints
-            @inbounds biomasses[t] = @views mean((1 .- images[:,:,t]) .* masks[:,:,t])
+            @inbounds biomasses[t] = @views Float64(mean((1 .- images[:,:,t]) .* masks[:,:,t]))
         end
     end
     output_images!(images, masks, overlay, OD_images, output_dir, filename)
@@ -315,9 +315,9 @@ function image_processing(image, blockDiameter, fixed_thresh, sig, output_dir, f
     if Imin != nothing && Imax != nothing
         OD_image = Array{Gray{Float32}, 2}(undef, size(image))
         OD_image .= (-1 .* log10.((image .- Imin) ./ (Imax .- Imin)))
-        biomass = mean(OD_image .* mask)
+        biomass = Float64(mean(OD_image .* mask))
     else
-        biomass = mean((1 .- image) .* mask)
+        biomass = Float64(mean((1 .- image) .* mask))
     end
     output_images!(image, mask, overlay, OD_image, output_dir, filename)
     return biomass
